Additional tests for the account service class

+++ bosco_nzeyi_code/project0/mockBank0/src/main/java/com/revature/mockbank/services/AccountService.java
@@ -89,7 +89,7 @@ public class AccountService {
             return false;
         }
         try {
-            int d = Integer.parseInt(strNum);
+            double d = Double.parseDouble(strNum);
         } catch (NumberFormatException nfe) {
             return false;
         }
